[Support] Add initial minimal bindings to NVTX (#25685)
This is to set the stage for us to plug it into the profiler (via a
CMAKE option). Note that this does not link the nvtx library into any of
our code and everything is done soley through a dlopen mechanism.

MODULAR_ORIG_COMMIT_REV_ID: ca1205d7f0cdb9d35777dadc35b06a209d87681b