[LLDB] Fix handling of unparsable DWARF types, pointers and kgen.invalid (#25704)
I had made a wrong assumption on how LLDB works because the Go debugger
tricked me when I was using it as inspiration. It turns out that thing
is flawled...
Anyway, LLDB expects all types, even the unparsable ones, to produce a
`lldb_private::Type`, even if it holds an invalid opaque compiler type.
So I'm now creating that object for types that failed parsing.

This was surfaced by a double pointer to a very unique type used by the
FFI API:
```
  tag = DW_TAG_unspecified_type, name = '!llvm.func<ptr<struct<()>> ()>'
```
In order to handle this type, I had to add support for
`DW_TAG_unspecified_type` and change the way we handle pointers. LLDB
expects to resolve pointers on its own and not by the dwarf parser. For
this type that failed to be parsed, LLDB creates a `void **` to its
address. Yes, `void **`, because LLDB=clang, but at least it doesn't
crash. I'm creating an issue to eventually display this as a
`pointer<pointer<ui8>` or something like that.

Besides that, I'm adding a small fix for dtype.invalid being emitted
with name 'void', and there's an issue to track the proper fix for that.

This fixes [Internal link] and now I
can see stack traces and variables.

MODULAR_ORIG_COMMIT_REV_ID: 18b77e957f64fe2324b10978e382155034bd60ca