[KGEN][Mojo] Remove risky `TypeConstantAttr::get` overload (#25804)
An overload existed that would default the metatype to `AnyRegTypeType`.
This patch removes that so that users are forced to propagate the
metatype. This paves the way to decoupling `AnyRegTypeType` from kgen,
and be able to eventually move it to `lit`.

MODULAR_ORIG_COMMIT_REV_ID: 1f0fff80a072297d5a2e31b2ea6c4fc29d3e9a32