[mojo-lang] Make conformance checking work for register passable types (#25790)
This involves transforming the signature to the register-passable
version and matching against that in the structs. Of special note is
the fact that trivial types are not allowed to define `__copyinit__` or
`__del__`, so they are specially handled.

This doesn't implement any thunk synthesis yet.

MODULAR_ORIG_COMMIT_REV_ID: fd608aec39fa422ef5b3a7e7244a301e6b086d51