[mojo-lsp] Add initial support for code blocks within doc strings (#25528)
This PR builds out initial support for handling code blocks within
documentation strings. This PR treats the code blocks within a
single doc string in a similar way as a jupyter notebook document,
where each code block is treated as a separate cell. The code blocks
are fed to the parser as REPL expressions disjoint from the main
document, this allows for reuse of state (such as imports), but also
allows for the code blocks to be evaluated in isolation.

This handles a majority of the text document requests fairly well already,
such as diagnostics/hover/definition/etc., but a followup is needed to
plumb through support for code completion and signature help.

See #20156

MODULAR_ORIG_COMMIT_REV_ID: 104f17ad28e1ae95577585d8cd6f2820a3e68ec3