[CMake] [KGEN] Fix idempotency violations due to DEPFILE misuse in KGEN (#25723)
This was a regression from #24991. That change sought made it so that
the MOGG mlirbc file would be rebuilt if the kernels changed, but due to
wonkiness about `DEPFILE` interpretation, it actually made those targets
always dirty, breaking idempotency (#19635). We use the same workaround
we use in many other places, making the output path relative to make
Ninja happy with the `DEPFILE`. After this, the build becomes idempotent
again. I am going to create a separate PR for an idempotency enforcement
mechanism.

In the future, we can clean this up once CMP0116 can be set to NEW, but
upstream LLVM needs to be adapted to handle CMP0116 before we can start
doing anything there.

MODULAR_ORIG_COMMIT_REV_ID: 6dd52c48ed090ccf6eb4e0ccbc3136d9686593bc