[mojo-lsp] Add support for semantic tokens (#25658)
This allows for IDEs to provide better highlighting for
symbols that aren't statically analyzable (is this a identifier
type/variable/module/etc?). Most of this PR is just plumbing
the structure through, which is kind of weird because of how
the LSP spec structures semantic tokens. The general structure
is inspired by how clangd implements this.

MODULAR_ORIG_COMMIT_REV_ID: 964dbe370038378076c8c5e74ef75cf8bd167a98