[LLCL] [Mac] Default to performance cores for num_threads=0 (#25919)
Instead of hitting a mix of performance cores and efficiency cores,
limit the number of threads to the number of local performance cores.

See #25659 for more information about this problem.

This is also made more useful in light of #25678, which will allow more
complex default logic at the lower levels of the stack.

Without this change, the out-of-the-box performance on my M2 pro on the
Python Bert sample is:

```
0.014743s
0.015750s
0.015330s
0.011771s
0.013818s
0.016406s
0.015669s
0.015948s
0.014969s
0.013153s
0.015646s
0.017820s
0.014421s
0.013060s
0.013182s
0.016315s
0.014938s
0.013384s
0.013708s
0.015352s
0.015067s
0.013837s
...
```

With this change, the out-of-the-box performance is (lower is better,
this is time):
```
0.007796s
0.007602s
0.007680s
0.007694s
0.008142s
0.007925s
0.007741s
0.007705s
0.007791s
0.008107s
0.007888s
0.007696s
0.007643s
0.007734s
0.008070s
0.007788s
0.007838s
0.007684s
0.007685s
0.008151s
0.008121s
0.007663s
0.007612s
0.007734s
0.008080s
0.007906s
0.007755s
...
```

MODULAR_ORIG_COMMIT_REV_ID: f62b4997a9ffc2ba4725f86f4dd951369c2110e7