[KGEN] Enable rebind to passthrough vtable changes (#25955)
This is a hack. The real problem here is that `rebind` is the wrong op
to be using here. We should introduce a separate op for augmenting type
information and use that instead, since this is altering the semantics
of `rebind`.

MODULAR_ORIG_COMMIT_REV_ID: 752b7d32218deb2ef5193a002e0ad0c147e098bf