[KGEN] Don't store the metatype in PointerType (#26056)
There is no need to propagate the metatype in the pointer type, since
the metatype just needs to be kept in the element type, for which there
is already a mechanism in DeclRefType. This means we can get rid of a
bunch of rebinds emitted by the parser that use rebind in a way that it
should not be. That also means we can unwind the rebind hacks that were
previously introduced.

This also means result types from calling into generic functions are
propagated correctly, so add a test for that.

Fixes #26035
Fixes #25916

MODULAR_ORIG_COMMIT_REV_ID: e3a4af0292446348e3381c30602453b534abb4c9