[KGEN] Extend compile_assembly poc to generate LLVM assembly (#26071)
This is useful when debugging to understand what LLVM code is generated
and how it maps to the generated assembly. Currently, one has to go in a
weird way to do this, and it's even more convoluted when separable
compilation is at play.

Note that this does not plumb this all the way to actually generate llvm
code, but that will be fixed in a subsequent PR

Part of #25373

MODULAR_ORIG_COMMIT_REV_ID: 1b81b189e6ae47affc4a3affa6111625f5dfe8b9