[KGEN] Propagate the emission kind correctly to the elaborator (#26102)
We were overriding the emission kind to be always asm, but that's not
correct. The fact that we were not honoring the emission kind was only
observed on the mojo code side of things.

MODULAR_ORIG_COMMIT_REV_ID: 08de7b150ef19d3e1924c90c695a040f1a42face