[Support] Add `open` and `refresh` flows to EntitlementStore. (#24652)
This patch adds `open` and `refresh` flows to the EntitlementStore. This will allow us to ping Phoenix to bootstrap (for `open`) and refresh (`refresh`) the client certificates. The very basic functionality is tested in unit tests, just so any readers can form a complete picture of the functionality. There is of course additional work going forward, but this should outline the majority of it.

The thing that isn't yet implemented is client key rotation - we could implement that in `refresh` or we could add it as a separate API. Given that we don't have a way (or plan to, in the near future) for a client to initiate a key rotation without simply deleting the key and certificate, we can wait on that.

MODULAR_ORIG_COMMIT_REV_ID: c0007cb856affba7e7238123b7e32926e16efb4c