[Profiling][Perf][Analysis] Add ";" seperator between manged name and details (#26134)
The operator details is a ";" separated list of "key=value" pairs. In
[Internal link] the manged operator
name was replaced with an index and stuck in the details instead, but
the separator was omitted. This presented a challenge for downstream
scripts that tried to parse the details.

This commit simply adds that seperator back.

MODULAR_ORIG_COMMIT_REV_ID: 1d6c7865a4d779f564f4a983cc3e25d015c56516