[mojo-parser] Use top-non package dir when resolving imports (#26216)
This allows for modules within a package to import modules/packages
at the same level as the package itself. In addition to making it easier to
compose packages/modules in a directory, this also makes it trivial
for doc tests within a package to import the package itself.  This
realistically shouldn't create any ambiguities because we already require
that modules within the same package be imported using `.`.

MODULAR_ORIG_COMMIT_REV_ID: ba623088ff8cc2a169c1dd93fcecc49fb33ec954