[MOGG] Slice out fusion attributes from new API (#25863)
Adds the infrastructure to both extract the attributes from the new API
and annotate at the kgen level and the logic to consume them in MOGG to
attach them as usable fusion hooks. The slicing will annotate the input
/ output lamdbas, detect if the kernel is elementwise and then attach
these attributes to the kernel.

MOGG picks them up and preserves them for the fusion passes to look at.

This allows them to be consumed by the fusion passes.

MOGG_ORIG_REV_ID: b47677ab1e2cc9161a01f6bbcd32226689534524