[Support][DebugInfo] Add DIExpression to ValueOp (#26254)
DIExpression PR2: Add DIExpression in ValueOp

Due to the default-constructed DIExpression in ValueOp, every ValueOp
naturally comes with a typed DIExpression. This requires users to update
the type of the expression when the type of the local-variable changes.
So far only one place triggers this, so this PR also changes that in
KGEN, but does not actually enable any functionality.

---
DIExpression PRs:
1. [Internal link]
2. ➡️ [Internal link]
3. [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 7190f6f6901165110cd6cd540917c6be42f89192