[All][mojo] Re-enable LLDB build on Windows (#22450)
LLDB should now support the necessary build configs
necessary for us to build and use it on windows. Lots
of functionality is still missing, but this will let us progressively
build that out. This also allows for removing all of the logic in
the build working around the fact that lldb may or may not be
built.

MODULAR_ORIG_COMMIT_REV_ID: 9286e57024316cd666fbf7f38be57b21ab86d399