[mojo-stdlib] Add StringTrait (#26414)
Add a trait for when a type implements the `__str__(self) -> String` function.
There's also a raising equivalent.

Apply it in the vocabulary types in the stdlib showing they conform to the
trait, along with the tests in the stdlib.

Note:
- Usages outside the stdlib will get updated in a separate PR
- Some vocabulary types in the stdlib conceptually satisfy the string trait but
  can't opt in to it yet.  For example, there was a use with `__str__(self)`
  returning a `StringRef` in one case, and `StringLiteral` in another case.
  This is another case of where the "convertible to String" would be a good
  idea I think.
GPU_TEST_ORIG_REV_ID: 909aaadc40b3d446f45add2cd4acfa607b887177