[mojo-lang] Fix Issue #26410 (#26432)
This Fixes #26410 which is a crash in check lifetimes, because the
`__del__` method synthesized for register passable values takes
self in memory instead of in registers.  It doesn't make sense for
StructEmitter to reimplement the logic that already exists in
DeclResolution (the logic in synthesizeEmptyDtor is correct) but
there is a bit of a tangle here with how closure emission wants
to work.  For now, just avoid the issue.

MODULAR_ORIG_COMMIT_REV_ID: 4297473e988b62af38fb147703115beb97a04743