[vscode] Update validation for debug json schema (#25939)
- Remove `program` as required attribute for `launch`. Now we support
both `program` and `mojoFile`, so this prevents a warning message from
appearing on the screen when using `mojoFile`. I also tried using a
`oneOf(program, mojoFile)` directive, but it show useful messages and
they sadly can't be customized, so I'm going for a simpler approach.
- Add the `unevaluatedProperties` directive, which shows a warning
whenever you have an entry in the debug config that is not defined in
the schema. This can help reduce typos.

MODULAR_ORIG_COMMIT_REV_ID: 383c9489eed8f4751ae67969063faf3568c7b598