[mojo-lang] Don't crash if parameter value emission fails (#26530)
We don't check in the implicit conversion whether a struct conforms to a
trait and just assume that it does if it specifies the trait as a
parent, since it's part of the signature.

That means the covnersion emission can fail if it does not conform.
Don't crash when that happens.

MODULAR_ORIG_COMMIT_REV_ID: 6a2a0b4907fb86256baf1ef63a84f6367c4bb7d0