[Support] Faster profiling (#26466)
> LOW PRIORITY, not to get in the way of modcon traffic!
> Though I did do this while procrastinating on my own modcon tasks :-|

Try to speed up profiling:
- Emplace the profiling entries directly as they are created, rather
than creating-then-moving them when they are recorded. This means when a
profiling entry is created it returns an event id which is later used to
end the entry.
- Avoid string copies when creating a profiling entry. This has four
aspects:
- `StringLiterals` can be recorded as is with no copies until the final
profiling output is being produced.
- Introduce an `InternableString` type to represent long-lived
`StringRefs`. Eg for those which come from `MEFAttr<StringRef>`. These
can be recorded as is, but a `TimeTraceProfiler::intern()` method is
supplied to ensure they are copied before the underlying storage is
freed. Eg before a `MEFFile` is to be freed. This copying cost can be
paid outside of the critical path.
- It is very common to need to pair a string with an integer value.
Rather than require string manipulation, allow that combination to be
recorded directly.
- Finally, rather than constructing the name of a 'nested' profiling
entry at create time, instead allow the entry to capture just the event
id of the parent event, and deal with all the name prefixing when the
profile is being written. Use this for the profiling entries for
sub-tasks in `KGEN::OutputChain::executeAsTask`, which is now much
simpler.
 - Remove unnecessary cruft from the API.

With this, using bert-base-uncased-seqlen-128 with all profiling
disabled as a baseline, I measure a ~5% perf hit for the default
profiling level but no --time-profile flag, and a ~10% perf hit with the
--time-profile flag.

Note my C++ fu is too weak to figure out how to support writing:
```
  ProfilerEntry<true>::create("a.literal.string")
```
which must instead be written as:
```
  ProfilerEntry<true>::create(StringLiteral("a.literal.string"))
```
This is *very* annoying since that's the common case. Most changes are
just doing that.

MODULAR_ORIG_COMMIT_REV_ID: 627de403a47038801a6703f21f1b5b39a659e1b3