[Mojo] Disallow default values of `inout` arguments (#26604)
Per the deep dive discussion on Nov 15, 2023 (see notes:
[Internal link]
we decided that implicit RValue to LValue conversions should not be
allowed. A corollary of this is that default values on `inout` arguments
are questionable, and until a use case arises that absolutely
necessitates it, we should disallow it.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: 40b2342585c012a95c594fe859ffcf332a4988c6