[mojo-tooling] Compute nested decl views only when necessary (#26595)
We create views for a lot of different purposes, and currently only
the json generation actually wants to know all of the nested views.
This PR moves the computation of views to only when we actually want
them, which provides a dramatic speed up for other tools (e.g. when
hovering over a module name, we don't need to generate a massive view
anymore). If we need the nested views at some point in the future, we
can add accessors to compute them, for now they've been moved to just
json generation.

MODULAR_ORIG_COMMIT_REV_ID: 2b1a34061b7471c81516fdc7b6122766c5477cf7