[Kernels][GPU] fix edge case in elementwise kernel (#26640)
Addresses a bug in the elementwise GPU kernel which is triggered when
the input buffer is non-contiguous and the last dimension is not evenly
divisibly by simd_width. This results in the simd_load/store methods
under-indexing into regions of memory which are adjacent on disk but not
adjacent in the tensor view we're operating on.

ALGORITHM_ORIG_REV_ID: 1d36726780a3a09549609814387a7edf37be04d5