[LLDB] Add visualizers for DynamicVector (#26723)
- Adds the synthetic children provided needed to print each element of
the vector. This works both for DWARF and REPL.
- Adds a summary provider that is used to render a short description of
the vector. It makes rendering of inlined variables on vscode nicer. It
shows the number of elements and, if the children are easy to serialize,
they are appended in the summary string. Sadly it doesn't work on the
REPL because my summary provider for decorator-based types conflicts
with this provider due to a bug in LLDB (if you have a
DynamicVector[Int], it'll show the first element as an index type and
the second element as an Int type, which is weird...). I spent a few
hours trying to chase this down but I gave up... I created an issue
anyway.

---------

Co-authored-by: Brian Gesiak <brian@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: de13f467af45ad33ef6c2eb6a56234a5a18f5dcf