[ONNX] Threading GPU device through Runtime (#26639)
[Joint with @jakubtuchol and @mbs-modular]

We realized that we weren't plumbing the GPU device information through
the runtime. This PR allows for us to get those options into the runtime
and propagate them down to the compilation process. We've verified that
this allows for the `llama2` demo to work on GPU.

---------

Co-authored-by: Jakub Tucholski <jakub@modular.com>
MAX_PYTHON_ORIG_REV_ID: 54551c4971a9b10e34c21f5983a083d04cf3b6be