[KGEN][DebugInfo] Track value changes during mem2reg (#26165)
Fixes #23020, Fixes #22592.

Need to track changes to a variable during mem2reg pass.

When a stack alloc gets promoted, every store to the original memory is
an update to the value. To track changes to the value, we create one
debug ValueOp for the initial undef, and one debug ValueOp for each
original store.

As a side effect of this change, a single source level variable may
correspond to multiple debug ValueOps. This makes the original Debug
Value -> Debug Declare conversion no longer straightforward. We may want
to consider rethinking that conversion in the future, but for now this
PR simply limits that conversion to ValueOps referencing unique source
variables.

Another side effect is that the debug ValueOp of a source variable may
be inside nested scopes (e.g. inside a lexical block). This means the
existing invariant that the loc scope and the valueInfo scope of a
ValueOp must be identical is too strict. We actually only need to make
sure the valueInfo scope is an ancestor scope of the loc scope. This way
we allow modifying values inside nested scopes.

---
Pointer-related DebugInfo PRs:
1. [Internal link]
2. [Internal link]
3. [Internal link]
4. ➡️ [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 8c0435fabfab7ca9ffd9446323cf899f63ba8855