[Mojo][NFC] Add `check-mojo-builtins` target (#26710)
This is a new target that includes parser tests that rely on builtins.
Ideally, we want to keep these tests to minimum: it is better to be able
to test the parser in isolation. However, this is not always possible,
so this new test suite will allow us to decouple _most_ parser tests
from the builtins.

MODULAR_ORIG_COMMIT_REV_ID: c5dba23564e9923663e174bb9f5d54e4d4f2afaa