[Support] Add alias mode to Buffer. (#26914)
This patch adds a way to have a read-only alias of a BufferRef. This is useful for cases when we may write lots of data to a WriteableBufferRef, and then want to return BufferRef aliases to parts of that data.

MODULAR_ORIG_COMMIT_REV_ID: 1851015bae5c952e8da7c6a21c07f04fc90f8450