[Kernels] Implements neon dotprod matmul (#26789)
Graviton2 and the M1 have the neon dot product instruction. This is more
or less exactly the same as the x86 VNNI instructions. Except VNNI is
only for u8s8 and the NEON dotproduct instructions are only for u8u8 and
s8s8.

The kernel used in this cases is based on the VNNI kernel. A more
optimal solution would be based on the NEON fp32 matmul method.

![Top model shapes matmul Gravition2 1 thread
[Internal link]

PR/main matmul performance.

[Internal link]

LINALG_ORIG_REV_ID: 9858cead898d30aa55567600137f294c71017acb