[CUDA] Preliminary support for pinned host buffers (#26948)
This allows allocating host buffers which are pinned for efficient
`cuda.host_to_device` and `cuda.device_to_host` transfers. Note that
CUDA requires such buffers to be allocated/freed with
`cuMemAllocHost`/`cuMemFreeHost`, and it is not sufficient to simply
`mlock` an already `malloc`ed buffer.

Though the resulting transfers are practically no-ops:
 - The `cuMemAllocHost` calls can take many milliseconds (!)
- There seems to be an execution stall after the transfers which I don't
understand.

So this is all off by default, but can be turned on for experimentation
with something like `--min-pinned-host-buffer-size=65536`.

I think our effort is better spent:
  - allowing shared CUDA buffers to be passed in to execute
  - allowing execute to be called in DPS style
  - supporting rewrites to iterative models such as llama2 so
    that the cv_cache can be updated in place instead of returned
    in concatenated form
  - supporting static memory planning for cuda so the
    pinned alloc could be done at init time

Changes:
- Three new ops `cuda.cpu.buffer.alloc.{static,dynamic,spec}`, to mirror
the regular `mgp.cpu.buffer.alloc.*` ops, but will use `cuMemAllocHost`
on the appropriate CUDA runtime.
- A new pass `CUDAPinHostBuffersPass` to replace regular with pinned
allocations when the host buffer is involved in a CUDA transfer. Note
that I try to allow the allocations to be done as early as possible
given their latency.
 - Primitives for the above.
- Since we want the pinned buffers to look just like regular CPU
buffers, but they must be freed with `cuMemFreeHost` against the
matching CUDA runtime, use the 'foreign' buffer facility of
`GML::BufferRefs` to track the extra context data.

MODULAR_ORIG_COMMIT_REV_ID: d98cd50e1651e427586072f41ad3e50c5091bf53