[Cache] Add multi-find capability. (#26915)
This patch adds the ability for backends to perform multi-finds in a batch. The default implementation is not interesting, but this enables backends (especially remote ones) to batch requests and perform compression transparently on a whole set of CAS objects. This allows a backend and server implementation to provide dynamic cross-blob compression on-demand.

MODULAR_ORIG_COMMIT_REV_ID: bb77de46e904bd16a2400fbd002f890c1f91df8d