[Support] Add `write` capability to mapped WriteableBuffers. (#26979)
This patch slightly modifies the storage for memory-mapped files so that if we open a file for writing with mmap, we can use `write` with a write pointer that starts at the beginning of the mapped file region. The write pointer gets incremented as we write to the buffer. This patch will be used for a behavior modification in the BlobCache that will enable us to fetch multiple CAS objects into a single buffer.

MODULAR_ORIG_COMMIT_REV_ID: 2b1611ca8d8944ab08a33b28ecfba27ceba591e4