[mojo][DebugInfo] Emit bools with 1 byte size (#26846)
The most frequent crashes I've seen while debugging are due to bools
having 0 byte size. In fact, when we create the debug info, we are
assigning it 1 bit of width and 1 bit of alignment, which eventually
gets converted into 0 byte size. But in any case, 1 bit seems to be
wrong.

A comment in DType.h says:

```
// kBool != ui1.  Like it, this only contains 1-bit of data, but it occupies
// 1-byte of storage.  The rest of the byte is guaranteed to be zeros.
```

This means that in terms of debug info, we care about the effective
storage size, which is 1 byte.

Besides that, we have this method that converts DType to MLIRType that
seems to be wrong

```
IntegerType M::getEquivalentIntegerType(MLIRContext *ctx, DType dtype) {
  if (dtype.isBool())
    return IntegerType::get(ctx, 1, IntegerType::Signless);
```

Shouldn't the width in this case be 8 as well instead of 1?

MODULAR_ORIG_COMMIT_REV_ID: f3943a92d26937d7cbb367520746066e9f45ed0a