[KGEN] LowerPOPToLLVM: Build hoisted allocs with parent region location (#26832)
Motivated by #26609 (partial fix).

In LowerPOPToLLVM, don't build alloc ops at the front of the block. This
spreads code from one source file line to multiple disjoint places in
the function and can be confusing for the debugger experience.
Instead, build the hoisted ops with a fused location of
both the original alloc and the parent op. This way it's clear that this
op was derived from the original alloc, but has a new meaning that's
related to the parent function.

MODULAR_ORIG_COMMIT_REV_ID: ad89df5492afc11ae4ad3db6a0d6b97e12022aa1