[mojo-lang] Warn about some dubious direct calls to functions. (#27274)
This enhances the Mojo compiler to warn about two patterns:
1) Directly self recursive function calls.
2) Calls to del on non-RValues, which copy a value and then delete the
copy
    (not the original value).

The former pattern was requested in
[Internal link]
but happens in our code as well:
[Internal link]
This patch removes those two functions to unblock progress.

The later was happening a lot when we were moving onto automatic
destructors and were migrating code, but was easy to catch with
the other infra.

This fixes [Internal link] and
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 3d58b9513e1522d78be0afa71ba1d7464c4042e2