[MO] Switch `mo.while` to not be `IsolatedFromAbove` (#27258)
This removes the `IsolatedFromAbove` interface from `mo.while`, and
simplifies the lowering to `grt.while`, deferring most of the work to
`fixupIsolatedFromAbove`. Also simplify the lowering of `mo.if` to
`grt.if` to similarly depend on `fixupIsolatedFromAbove`.

The upside is only those SSA values which must be 'updated' by the loop
body need be spelled out in the input parameter list, and code which
wants to just do syntactic use-def chasing can 'see' across `mo.while`
boundaries for the 'unchanged' values.

The downside is all `mo.while` operands must be 'renamed' using an `as`
clause. Eg:
```
  %x = ...
  %y = ...
  mo.while (%x as %xi: !foo) {
    ... %xi ...   // ok, 'current' version
    ... %y ...    // ok, outer version
    ... %x ...    // ok (**), outer version
  } do {
    %new_x = some.op(%xi)
    mo.yield %new_x
  }
```

(**) Note it's not possible to reject this since passes are free to
introduce references to outer values inside the while body independently
of the while signature.

While in there add the forgotten `HasOnlyGraphRegion` interface to
`mo.{while,if}`.

I've *not* updated:
 - ONNX/lib/Conversion/MONNXToMO/ControlFlowOps.cpp
 - PyTorch/lib/Conversion/TorchToMO/ControlFlowOps.cpp
 - TensorFlow/lib/Conversion/TFToMO/TFToMO.cpp

to remove the free-variable binding logic to exploit the relaxed
constraints. Some of that code does more than just free-var binding so
decided to leave alone.

MODULAR_ORIG_COMMIT_REV_ID: 5ecee975a5071ab2eb34aa96fbb5b0d985ecd661