[Telemetry] Tag Mojo Metrics with Filename (#27128)
Tagging our Mojo timer metrics with the the filename of the mojo we are
compiling. This is going to be very helpful when tracking mojo
compilation metrics and regressions in CI.

Note that a few of these metrics need to have a lower bound outlier
filtering to account for the cache hits. We can add an additional tag in
the future for those, but this should get us most of the way there.

Had to do a lot of type mangling to get OTel to accept these owned
values (it accepts this heterogeneous blob of variants of un-owned
values and I learned a lot of new things about std::visit which I've
never had to use before), and there are probably better ways to do some
of this, so if there is let me know for the sake of my learning!

MODULAR_ORIG_COMMIT_REV_ID: b698086805e2734a000360d1df74302700884342