[Mojo][NFC] Refactor `OverloadFitness` and candidate comparison logic (#27320)
The logic was encoded implcitly in two places: a goofy numeric payload
that used its lower 3 bits to contain boolean values used in a decision
tree, and a helper function that implicitly relied on this, but also
introduced some logic to break ties based on the number of parameters in
a signature. This patch completely refactors this so all the overload
fitness comparison logic now lives in `OverloadFitness` and all metrics
used for comparison are stored explicitly. This allows cleaning up the
call emission and overload evaluation logic a bit as well.

MODULAR_ORIG_COMMIT_REV_ID: ddf32d51598946a077c33da2c7d3b97ed9a2031c