[KGEN] Propagate MODULAR_RELEASE_PACKAGE_BUILD to mojo files, NFC (#27339)
This is useful to identify whether a mojo file is built to target a
release or not. This will be used in subsequent PRs to feature gate some
functionality if we are bundling the mojo files into the product.

MODULAR_ORIG_COMMIT_REV_ID: b39ef9e3581c9d4ac316ae4a26ff0d1f95d41908