[mojo-lsp] Don't add a def range for modules (#27298)
Modules have their location set to the beginning of the
document, but aren't actually defined in the document. This PR
ensures we don't treat the start of the document as a reference to
the module (which also fixes some really weird semantic token issues
when importing the main module in a doc string).

MODULAR_ORIG_COMMIT_REV_ID: da4dddc7dff04b7f90f98f05cb2699f0b598239d