[mojo-repl] Add support for a '%cd' magic (#27360)
This is a limited version of the builtin ipython magic that lets
users change the current working directory, or switch to a previous
directory. Given this is implemented in the REPL, this also works
naturally in jupyter notebooks as well.

MODULAR_ORIG_COMMIT_REV_ID: 74e1411a1665105ee6af6256e831545de2f77a8c