[Support] Support certificate chains for client certificates. (#27209)
This patch enables Phoenix to send a full chain of certificates - this is important because we will likely want a realistic chain like this:
```
root -> intermediate -> intermediate -> client
```
where only the root certificate is embedded in the binary. That means we need to staple (concatenate) the two intermediate certs to the client cert when returning it from Phoenix so that the client can construct a valid chain of trust.

This patch abstracts the certificate chain into its own class, and adds assertions that the various usages of said certificate chain verify the certificate is valid before actually trying to use it for something.

Finally, since I was editing the whole file anyway, I've moved the EntitlementStore methods to match the order in the header file.

MODULAR_ORIG_COMMIT_REV_ID: bae215f6111f95d5774df1ef42ceac73782297bd