[Support] Update CertificateChain to take a CRL buffer. (#27210)
This patch makes a minor change that has CertificateChain::verify take a CRL PEM buffer. This makes it the responsibility of the EntitlementStore (which already has an HTTPClient) to fetch and cache the CRL on the system. That change is still in progress, but this should clarify the boundaries between what is the EntitlementStore and what is the CertificateChain.

MODULAR_ORIG_COMMIT_REV_ID: 6b879bc7b4556c7af676773a3c65546df436fcd6