[Support] Rotate client keypairs on cert refresh. (#27241)
We can generate a new key every time we refresh the certificate. In order to make this work, we want to sign the CSR with the old key (whose associated cert is still valid) and add the signature so that Phoenix knows that the CSR is coming from the expected client.

MODULAR_ORIG_COMMIT_REV_ID: e23c85db26b4bb9f500863953b304941a0d7545d