[Support] Add CRL checking and fetching. (#27315)
This patch adds CRL checking to the certificate verification. It also adds fetching of the CRL from `https://crl.modular.com`. On a failure, we attempt to parse it from the filesystem. If that also fails, we simply fall back to not having a CRL. This is not particularly aggressive, but it's the only way to support a use case where the client is offline from the first run. It's also generally implementation-defined to establish what to do if the CRL cannot be found, and this kind of behavior (ignore it) is pretty common.

MODULAR_ORIG_COMMIT_REV_ID: f2c1530b5fc7cabce1702c2aa643d2c53be77e6b