[Models] Add Defensive checks to `llvm::alignTo` calls (#27356)
We have been regualarly been hitting a hard to debug crash in model coverage and verification CI.
The issue is track in #26118.

Fundamentally, there is a location where a 0 alignment is passed into `llvm::alignTo` and an assert fails.
This bug is exceptionally rare and only has been reproduced in CI.

Despite debugging in CI with ssh sessions, the issue has not been figured out.
The issue does not produce a useful stacktrace.
The current assumption is that the failing call comes from a jitted kernel that does not have debug info.

These sorts of defensive checks were added to two PRs (#26610 and #26774).
After more than 40 runs of the PRs collectively, the issue has not arisen.
Due to the hard to debug nature of this problem, we are adding the checks to the actually repo until the bug is pinpointed.

Unless I missed a use of `llvm::alignTo` all callsites that aren't obviously nonzero now have a defensive check.

MODULAR_ORIG_COMMIT_REV_ID: 4be721fb8f07c1a88a82d2639ab03c4801d5738f