[KGEN] Translate LineTables into DebugDirectives for NVPTX (#27385)
Addresses point 4 of #26807.

I hate it... but the NVPTX backend requires debug level of
DebugDirectivesOnly in order to have line information & not drop
performance. LineTables is treated the same as FullDebug, which disables
optimizations (and cannot be turned back on). However, for host code,
which needs to generate into a binary, we have to use LineTables,
otherwise we don't get the proper dwarf information to recreate line
tables for stepping.

Ideally this hack is only for the short term, and we do something to the
NVPTX backend to make things more general.

Hack tracked by #27478.

MODULAR_ORIG_COMMIT_REV_ID: 8daa29d89a757f5d1f68d6bfe4defdc7630786fb