[LLCL][NOT FOR REVIEW] Support notion of 'current runtime', and expose it to Mojo (#27437)
[NOT READY FOR REVIEW. Want some bake time in ci first.]

Currently, we rely on the KGEN::OutputChain to convey context down into
Mojo parallelize, including on which runtime the sub-tasks should be
launched into. We want to get rid of that, and instead allow parallelize
to just use the Mojo 'global runtime'. However, when Mojo is used via
the C API, the Runtime has already been established and used
pervasively. Allow that Runtime to be picked up.

 - Cleanup the LLCL globals handling.
- Introduce LLCL::CompactRuntimePtr::{get,set}CurrentRuntime so the
'owning' runtime can be associated with a thread.
- Introduce WorkQueue::associateRuntime to tell a work queue which
runtime to associated with its threads.
- Introduce KGEN_CompilerRT_LLCL_GetCurrentRuntime to return the current
runtime associated with the callers thread.
 - Update the Mojo _get_global_runtime to use the above when available.

In this way we can continue to have multiple concurrently-active
Runtimes (which is already baked into the C API), but avoid Mojo having
to be exposed to any of that complexity.

(For the record I think we should just have one global Runtime, but one
step at a time.)

MODULAR_ORIG_COMMIT_REV_ID: 6d17e196f9af7ad40e49bc12ee09d75f0cd7fbed