[KGEN] Fix alignment issue (hopefully) (#27574)
I am not 100% sure this is the root cause of #26118, but it seems highly likely.

When a vector has a size of zero or an element type with an alignment of zero, it will result in an output alignment of zero.
Alignment is never allowed to be zero. As such, a size of zero should return an alignment of one.

MODULAR_ORIG_COMMIT_REV_ID: 44754de5424a2bda705361a2f82158955d25172c