[Kernels][Profiling] Decouple tracing from out_chain (#27569)
This gets us one step closer to removing the out_chain.

Previously tracing was done on the out_chain. The out_chain was used to
maintain a connection from the host task to the async subtasks.

Here, we allow Trace objects to be constructed with a parent_id, which
will prepend the parent Trace name to the Trace name. The current parent
can be queried by calling `get_current_id`. This is used to connect the
async subtask to the main trace event in async_parallelize.

All instances of `out_chain.trace` are replaced by `with Trace() as t:`
and out_chain.trace() is removed.

Traces cannot currently cross async/await boundaries since they use the
implicit per-thread profiling stack.

In the future the kernels could have a `@trace` decorator and the `with
Trace() as t:` scope can be synthesized.

Co-authored-by: Ian Tramble <ian@modular.com>
NN_ORIG_REV_ID: 0997e11492b13664075d92f0534dda1d87e70dce