[LSP] Revive python tests (#27541)
This tries to revive the lsp python tests. This includes a workaround
for an issue that happens when mojo-lsp-server's stderr is too long,
which causes it to hang when trying to send more text to the test
client. I assume that the test client has some weird logic when
collecting stderr.

MODULAR_ORIG_COMMIT_REV_ID: 786fcb466999bfedd4b1471f894f9331fa4c80d7