[Support] Add docs and examples for writing CLIs (#27693)
This adds a "CLI Guide" doc that explains the tools we use at Modular
for writing command line interfaces.

The bulk of the documentation is actually in the form of an example CLI
that is also added by this patch: `greeter-cli`. The source code in the
example can be copy-pasted to quickly create a new CLI application at
Modular, and it contains code comments that explain what it's doing and
how to modify it.

(Note that unlike examples in the LLVM and MLIR project, this patch
doesn't add CMake options to include its targets or build it
conditionally, such as `LLVM_BUILD_EXAMPLES` -- the example is living
documentation, and it should be built in CI when all other targets are
built. I could add an option, but then I'd just update our CI workflows
to enable that option most of the time, which is silly.)

Fixes #16240.

MODULAR_ORIG_COMMIT_REV_ID: e947a918af154243647e3c0b310f44bf5bd81151