[mojo-lang] trait error instead of trip assert (#27510)
Before this PR, the `getSpecializedSignature` function's error emitter
argument was optional, but it was not supposed to error in cases where
none was given. The error emitter was made optional to make uses less
verbose in cases where no error was expected, but that also makes it
easier to not realize that you need an emitter and write code that is
invalid and can trip the assert.

This PR makes the error emitter mandatory again, but adds another form
that takes a location and builds and error emitter with it. Providing a
location is less verbose than providing an error emitter closure while
avoiding the issues of providing no error emitter.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 3b3553a8758edd5fefe1750b76bfad6d4b389980