[docs] Fix mojo changelog for hash function support (#27798)
[Internal Link]
was a little eager in updating the changelog as it mentions support for hashing
`String` and `Int, but that work has not landed yet.  It is open for review at
[Internal Link] but it will not make it in time
for the v0.6.1 release likely.

So, update the changelog to reflect the reality of the code today for the v0.6.1
release to avoid confusion for users.

modular-orig-commit: 4c6391a58b9176fafb68b3f54490818fdfec22ad