[MGP][MOGG] Poor man's Mojo 'async' support (#27803)
As we incrementally transition from `KGEN::OutputChain` to using Mojo
async/await we'd like to hit a 'safe' intermediate state were no
AsyncValue chains cross from the C++ runtime into Mojo. The approach is
to make all the existing Mojo kernels
'synchronous', ie:
- Fork/join parallelism is ok, but we will switch to a new `parallelize`
impl which uses `TaskGroup` on the default (ie C++ runtime-supplied)
`Runtime`. That will still use an `LLCL::await` internally on a local
hidden chain, which is a signal to us that the await needs to be hoisted
up into a Mojo-level await when we make the rewrite, which in tern will
trigger rewrites all the way up to the Mojo kernel entry points.
- Sequential composition is just Mojo sequential composition, with no
locally created `OutputChain`s etc.

However, the current `KGEN::OutputChain` approach is actually giving us
a big perf win in situations such as:
```
   %x = some.op ...
   %y = other.op %x
   %z = other.op %x
```
If `other.op` is an elementwise then currently both `%y` and `%z` will
be launched in parallel as soon as `%x` is ready. In our 'safe'
intermediate state the elementwise will be blocking, so to preserve the
current performance (we hope) we will need to wrap the `other.op` kernel
launch in an `addTask`.

Since we only want to do that for kernels which will eventually become
Mojo async functions, we introduce a temporary `@mogg_will_become_async`
decorator to signal when the `addTask` is needed. That also requires
some plumbing through MOGG to remember the difference between sync and
async kernels. However, some/most of that plumbing can be repurposed to
convey the same semantics, all be it derived from the Mojo kernel
signature rather than a hacky decorator.

REGISTER_ORIG_REV_ID: 252e23f5d24aef497bb82f1171b581844dfbcdad