[CI] Set a timeout to catch hanging tests (#27815)
Set a timeout on `llvm-lit` tests to get an idea of which test hung from
the CI logs, since it'd be annotated with "TIMEOUT".
Whether that test is actually buggy or not would still need
interpreting.
However as of today we don't even have the information about which
`llvm-lit` tests are still running when a GH workflow times out.

Only enable the timeout in CI with no sanitizers.
The timeout is currently a generous 60 minutes.

MODULAR_ORIG_COMMIT_REV_ID: 6348930b9f25c21b7b744c771d9f05dcbd91882b