[KGEN] Dump target and type info on align failure (#27568)
### [KGEN] Dump target and type info on align failure

This is continuation of work to track down #26118.

It seems that each failure is hitting this specific location, but we
want more debug info.
For this specific location also dump the type and target info to
hopefully see why the align is zero.

MODULAR_ORIG_COMMIT_REV_ID: 66e33374163057b328b6266de8c43305da971279