[mojo-lang] Don't track lifetime of `__get_address_as_uninit_lvalue` (#27673) (#27961)
This intrinsic returns an uninitialized pointer as an lvalue which
can be used in a variety of ways and does not need to be initialized
necessarily on all paths.  Just decline to lifetime track it, because
anyone using this has got to know what they're doing.

This fixes [Internal link]

This reapplies #27673 with testcase updates.

MODULAR_ORIG_COMMIT_REV_ID: abcb61cae84cd9b550cb8ffa8e782319fc090478