[mojo-stdlib] Fix crash in `DynamicVector` copy ctor (#27971)
In certain situations, `DynamicVector` copy constructor would incorrectly set
the internal `data` pointer.  For example, in the provided test case, it would
alloc and set the `data` pointer to be an empty (null) `AnyPointer[T]`.  Then,
it would immediately lead to a crash when a user tries to `push_back`.

Fix this behavior by doing the simple thing: delegating to the existing
constructor that accepts a capacity argument (which will correctly always set
the `data` pointer), and then add the arguments to the internal storage from the
copied-from vector.

Fixes [Internal Link]
Fixes [Internal Link]

modular-orig-commit: ecc8c55b98b27e7ce07ae2533770fb51de3e3f3a