[Support][Entitlements] Add `getUserID` and `alwaysOpen`. (#27824)
This patch supports the 'fail open' nature of #27787 by providing a method to always open an EntitlementStore, emitting a warning if we were unable to do so, but not actually hard-failing.

We also add the `getUserID` method in order to centralize the notion of 'identity' in our stack to the EntitlementStore. Future patches will adopt this method rather than checking `modular.cfg`.

MODULAR_ORIG_COMMIT_REV_ID: b03559ddedf3511fcaaaaaa6a93ffc2fe8ec75fe