[Build] [NFC] Update Support tests (#28002)
This updates the Support tests to include explicit cases and includes where
required (generally following IWYU, but not in a dogmatic way).

The test fails in the overlay since it currently depends on the custom fork of
googletest used within LLVM, which adds support for operation<< StringRef where
needed. This is a weak value proposition and ties us to this bespoke fork,
instead we can just make the test more portable (useable both with the
LLVM-fork of gtest, and a vanilla gtest) by adding a few casts.

MODULAR_ORIG_COMMIT_REV_ID: 86c04b2385ff6541f021fa0189750e873bba6238